Here is a much more recent list of Python Talks
Thank you for the wonderful Pandas library. 

For the growth of the Panda's ecosystem, it is very important to have a good collection of recent human-curated videos of technical talks, organized by subject category.  I have built and am continuing to improve and maintain this list.  

I apologize for deleting the older talks, but as a curator of technical videos, I have found that it is best to provide people with fresher content.  

Marc mentioned that a new website will be coming online soon, but in the meantime for a few weeks, this will be a small improvement to your website. 

Your comments on this pull request are most welcome.